% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c.scanonevar.R
\name{c.scanonevar}
\alias{c.scanonevar}
\title{c.scanonevar}
\usage{
\method{c}{scanonevar}(...)
}
\arguments{
\item{...}{the scanonevar objects with permutations to be combined}
}
\value{
a scanonevar object that is the concatenation of the inputted
scanonevars
}
\description{
combines scanonevar objects that have permutations to improve the precision of the p-value estimates.
}
