% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects_over_genome_plot.R
\name{effects_over_genome_plot}
\alias{effects_over_genome_plot}
\title{effects_over_genome_plot}
\usage{
effects_over_genome_plot(sov, effect.names = NULL, mean.or.var = c("both",
  "mean", "var"), se_ribbons = TRUE)
}
\arguments{
\item{sov}{the scanonevar}

\item{effect.names}{the effects to plot}

\item{mean.or.var}{mean or variance effects of covariates}

\item{se_ribbons}{Should a ribbon from estimate - se to estimate + se be plotted?}
}
\value{
the plot
}
\description{
Plots estimated effects and their standard errors at each locus in the genome.
}
