% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects_plot.R
\name{effects_plot}
\alias{effects_plot}
\title{effects_plot}
\usage{
effects_plot(sov, effect.names = NULL, mean.or.var = c("both", "mean",
  "var"))
}
\arguments{
\item{sov}{the scanonevar}

\item{effect.names}{the effects to plot}

\item{mean.or.var}{mean or variance effects of covariates}
}
\value{
the plot
}

