% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula.utils.R
\name{replace.markers.with.add.dom_}
\alias{replace.markers.with.add.dom_}
\title{replace.markers.with.add.dom_}
\usage{
replace.markers.with.add.dom_(cross, mean.formula, var.formula)
}
\arguments{
\item{cross}{The \code{cross}, built by \pkg{qtl} to be used in mapping}

\item{mean.formula}{The formula to describe the mean of the phenotype.
Keywords are mean.QTL.add and mean.QTL.dom for the additive and dominance
components of the QTL effect on the mean.}

\item{var.formula}{The formula to describe the residual variance of the
phenotype.  Keywords are var.QTL.add and var.QTL.dom for the additive and
dominance components of the QTL effect on residual phenotype variance.}
}

