% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.genoprobs.by.marker.name.R
\name{get.genoprobs.by.marker.name}
\alias{get.genoprobs.by.marker.name}
\title{@title Get Genotype Probabilities From a Cross Object By Marker (or Psuedomarker) Name}
\usage{
get.genoprobs.by.marker.name(cross, marker.name)
}
\description{
@author Robert Corty \email{rcorty@gmail.com}
}
\details{
@description \code{get.genoprobs.by.marker.name} is an accessor function that returns the
   genotype probabilities of each individual in the cross at the given marker or pseudomarker.

@param cross The cross from which the genetic information will be extracted.
 @param marker.name The name of the marker where we want to know each individuals most likely genotype.

@return Probability of each genotype at the given locus for each individual.

@details none
}

