% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.peaks.from.scanonevar.R
\name{get.peaks.from.scanonevar}
\alias{get.peaks.from.scanonevar}
\title{Get Local Maxima from Scanonevar}
\usage{
get.peaks.from.scanonevar(x, thresh)
}
\arguments{
\item{x}{the \code{scanonevar} object from which the peaks will be identified}

\item{thresh}{Optionally, the threshold over which a value has to be to be considered a peak.
For example if one locus has a LOD score of 1 and the loci to its sides have LOD score of
0.9, that's not really an interesting or "peaky" locus.  Defaults to 3 if x is in LOD units
and 0.05 if x is in p-values}
}
\value{
tbl_df of identified loci
}
\description{
\code{get.peaks.from.scanonevar} scans the genome for loci such that
  the locus to the left has a lower value and the locus to the right has a lower value.
  This value can be either LOD score or -log10(p-value)
}
\details{
none
}
\author{
Robert Corty \email{rcorty@gmail.com}
}

