% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanonevar_.R
\name{initizlize.scanonevar.result_}
\alias{initialize.scanonevar.result_}
\alias{initizlize.scanonevar.result_}
\title{initialize.scanonevar.result_}
\usage{
initialize.scanonevar.result_(loc.info.df, scan.types)
}
\arguments{
\item{loc.info.df}{the tbl_df that contains information about the
locs that are being scanned, minimally a 'loc.name', 'chr' (short for
'chromosome') and 'pos' (short for 'position') for each loc.}

\item{scan.types}{Type(s) of scan(s) to be conducted, one of: 'mean',
'var', or c('mean', 'var', 'joint')}
}
\value{
A tbl_df with the metadata for a scanonevar, once filled in by
the scanonevar function, it will be a scanonevar object.
}
\details{
This internal function should not typically be called by a user.
}
\examples{
x <- dplyr::data_frame(loc.name = c('m1', 'm2'),
                       chr = c(1, 1),
                       pos = c(5, 10))
initialize.scanonevar.result_(loc.info.df = x,
                              scan.types = 'mean')

}
\author{
Robert W. Corty
}

