% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.cross.R
\name{is.cross}
\alias{is.cross}
\title{is.cross}
\usage{
is.cross(x)
}
\arguments{
\item{x}{The object being tested for whether or not is is a cross.}
}
\value{
TRUE if x is a cross object, FALSE otherwise.
}
\examples{
is.cross(3)
is.cross(qtl::sim.cross(map = qtl::sim.map()))

}
\author{
Robert W. Corty \email{rcorty@gmail.com}
}

