% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.scanonevar.R
\name{is.scanonevar}
\alias{is.scanonevar}
\title{@title is.scanonevar}
\usage{
is.scanonevar(sov)
}
\description{
@author Robert Corty \email{rcorty@gmail.com}
}
\details{
@description Tests whether an object is a valid scanonevar object

@param sov the object which is tested for being a scanonevar object

@return Returns TRUE if 'sov' is a valid scanonevar object, with attribute 'why.not' an empty list
   Returns FALSE if 'sov' is not a valid scanonevar object with attribute 'why.not' a list of
   error messages for failed tests

@details none

@seealso \code{\link{scanonevar}}, \code{\link{scanonevar.to.p.values}}
}

