% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanonevar_.R
\name{LOD}
\alias{LOD}
\title{LOD}
\usage{
LOD(alt, null)
}
\arguments{
\item{alt}{alternative DGLM}

\item{null}{null DGLM}
}
\value{
log( likelihood of alternative model / likelihood of null model)
}
\examples{
x <- 27599
}
\author{
Robert W. Corty \email{rcorty@gmail.com}
}

