% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_var_plots.R
\name{sample_stats_mean_var_plot}
\alias{modeled_effects_mean_var_plot}
\alias{sample_stats_mean_var_plot}
\title{sample_stats_mean_var_plot}
\usage{
sample_stats_mean_var_plot(cross, phenotype.name, grouping.factor.names)

modeled_effects_mean_var_plot(cross, phenotype.name,
  focal.covariate.names = NULL, nuisance.covariate.names = NULL)
}
\arguments{
\item{cross}{the cross object from which phenotypes and genotypes are drawn}

\item{phenotype.name}{the name of the phenotype}

\item{grouping.factor.names}{the factors to group individuals by}

\item{focal.covariate.names}{the focal covariates, whose effects will be plotted.  Markers or phenotypes.}

\item{nuisance.covariate.names}{the nuisance covariates, whose effects will be modeled, then marginalized over.  Markers or phenotypes.}

\item{cross}{the cross}

\item{phenotype.name}{the name of the phenotype of interest}
}
\value{
Nothing, just plot.

nothing, just the plot.
}

