% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.scanonevar.R
\name{plot.scanonevar}
\alias{plot.scanonevar}
\title{plot.scanonevar}
\usage{
\method{plot}{scanonevar}(x, y = NULL, chrs = unique(x$chr), units.to.plot,
  col = c("black", "blue", "red", "forestgreen"), bandcol = "lightgray",
  legends = c("DGLM-joint", "DGLM-mean", "DGLM-var", "LM"),
  legend.pos = "top", legend.ncol = 2, legend.cex = 1, gap = 25,
  incl.markers = TRUE, title = attr(x, "pheno"), title.cex = 1.5,
  ylim = c(0, 1.05 * max(coords.y.locus, na.rm = TRUE)),
  show.equations = (length(chrs) != 1), alpha.side = "left",
  line.width = 1, vertical.bar = NA, ...)
}
\arguments{
\item{x}{the \code{scanonevar} object to be plotted}

\item{y}{Optionally, a \code{scanone} object to be plotting for comparison to the \code{scanonevar} object.}

\item{chrs}{Optionally, the subset of the chromosomes to plot}

\item{units.to.plot}{Optionally, whether to plot 'lods' or 'emp.ps'.}

\item{col}{Optionally, a vector specifying the colors of the scan lines.  Defaults to \code{c("black", "blue", "red", "darkgreen")}.}

\item{bandcol}{Optionally, a background color for the even-index chromosomes in this scan.}

\item{legends}{Optionally, the name to put for each scan in the legend.  Defaults to \code{c('mean or var', 'mean', 'var', 'scanone for comparison')}.}

\item{legend.pos}{Optionally, the corner/edge where the legend should be drawn.  Defaults to \code{'top'}.}

\item{gap}{Optionally, The space between chromosomes in Mb.  Defaults to 25.}

\item{incl.markers}{Optionally, whether to draw a rug plot along the bottom indicating where the markers are.  Defaults to TRUE.}

\item{ylim}{Optionally, the y limits for the plot.  Defaults to \code{c(0, 1.05 * highest.point)}.}

\item{show.equations}{Optionally, whether to write the modeling equations under the title.  Defaults to TRUE.}

\item{...}{Additionaly graphical parameters.}

\item{main}{Optionally, a title for the plot.  Defaults to the name of the phenotype from the scanonevar.}
}
\value{
Returns nothing.  Only makes the plot.
}
\description{
\code{plot.scanonevar} implements the plot generic for objects of class 'scanonevar'.
   Because scanonevar objects can be viewed in terms of LODs or empirical p-values,
   this plotting function checks the 'units' attribute to determine which to plot.
}
\details{


If such a strong signal was observed that the empirical p-value underflows R's
   float type, this function produces an error.  The author is open to suggestions on how
   to deal with this situation better.
}
\author{
Robert Corty \email{rcorty@gmail.com}
}
\seealso{
\code{\link{scanonevar}}, \code{\link{scanonevar.to.p.values}}
}

