% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_model_fitting.R
\name{pve}
\alias{pve}
\title{percent variance explained}
\usage{
pve(LOD, n)
}
\arguments{
\item{LOD}{the log odds between the null and alternative model}

\item{n}{the number of observations}
}
\value{
pve
}
\description{
percent variance explained
}
