% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanonevar.perm.R
\name{scanonevar.perm}
\alias{scanonevar.perm}
\title{@title Conduct Scanonevars on Permuted Genotype Data}
\usage{
scanonevar.perm(cross, mean.formula, var.formula, n.perms,
  chrs = unique(names(cross$geno)), verbose.return = FALSE)
}
\description{
@author Robert Corty \email{rcorty@gmail.com}
}
\details{
@description \code{scanonevar.perm} conducts many \code{scanonevar}s on permuted data
   and returns the maximum observed LOD score for each chromosome type in each scan.
   The results should be put into \code{convert.scanonevar.to.empirical.ps} with a scan in LODs
   to convert that scan to empirical p-values.  It's important that all the parameters used in
   \code{scanonevar.perm} are the same as the parameters that were used in the \code{scanonevar}
   that they will be used to convert to empirical ps.

@param n.perms the number of permutations to conduct
 @inheritParams scanonevar

@return Returns a tbl_df of maximum LOD score observed in each genome scan for each chromosome type.

@seealso  \code{\link{scanonevar}}, \code{\link{scanonevar.to.p.values}}

@details It is recommended to use approximately 1000 permuted scans to produce highly-replicable,
   publication-quality empirial p-values.  For this purpose, users are recommended to dispatch this
   function to many computers in parallel, carefully setting the seed on each computer to insure
   pseudo-randomness.

@details none

@examples
 \dontrun{
   my.perms <- scanonevar.perm(cross = my.cross,
                               n.perms = 10))

}
}

