% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/units.R
\name{units}
\alias{units}
\title{Get units of a scanonevar object}
\usage{
units(object)
}
\arguments{
\item{object}{the scanonevar object whose units are interrogated}
}
\value{
units of scanonevar object
}
\description{

}
\details{
none
}
\author{
Robert Corty \email{rcorty@gmail.com}
}

