% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.input.scanonevar.R
\name{validate.input.scanonevar}
\alias{validate.input.scanonevar}
\title{@title Validate the Input before Conducting a Scanonevar}
\usage{
validate.input.scanonevar(cross, mean.formula, var.formula,
  chrs = names(cross$geno))
}
\description{
@author Robert Corty \email{rcorty@gmail.com}
}
\details{
@description \code{validate.input.scanonevar} should not typically be called by
   a user.  It validates the input to the \code{scanonevar} function.

@param cross The cross on which scanonevar will be performed, to be validated
 @inheritParams scanonevar

@return Validated and organized inputs for scanonevar

@details none
}

