% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.scanonevar.input_.R
\name{validate.scanonevar.input_}
\alias{validate.scanonevar.input_}
\title{validate.scanonevar.input_}
\usage{
validate.scanonevar.input_(cross, mean.formula, var.formula, chrs,
  allow.no.qtl = FALSE)
}
\arguments{
\item{cross}{The \code{cross}, built by \pkg{qtl} to be used in mapping}

\item{mean.formula}{The formula to describe the mean of the phenotype.
Keywords are mean.QTL.add and mean.QTL.dom for the additive and dominance
components of the QTL effect on the mean.}

\item{var.formula}{The formula to describe the residual variance of the
phenotype.  Keywords are var.QTL.add and var.QTL.dom for the additive and
dominance components of the QTL effect on residual phenotype variance.}
}
\value{
TRUE is parameters are valid for scanonevar, errors with a
helpful error message if parameters are invalid.
}
\examples{
x <- 27599

}
\author{
Robert W. Corty \email{rcorty@gmail.com}
}

