% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanonevar.R
\name{wrangle.scanonevar.formulae_}
\alias{wrangle.scanonevar.formulae_}
\alias{wrangle.scanonevar.modeling.df_}
\title{wrangle.scanonevar.formulae_}
\usage{
wrangle.scanonevar.formulae_(cross, mean.formula, var.formula)

wrangle.scanonevar.modeling.df_(cross, scan.formulae, genoprobs)
}
\arguments{
\item{cross}{The \code{cross}, built by \pkg{qtl} to be used in mapping}

\item{mean.formula}{a mean.formula}

\item{var.formula}{The formula to describe the residual variance of the
phenotype.  Keywords are var.QTL.add and var.QTL.dom for the additive and
dominance components of the QTL effect on residual phenotype variance.}
}

