% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanonevar.R, R/utils_formulae.R, R/utils_stats.R, R/utilts_genetic_data.R
\name{wrangle.scanonevar.input_}
\alias{LOD}
\alias{additive.component}
\alias{dominance.component}
\alias{is.formulae}
\alias{is.mean.formula}
\alias{is.var.formula}
\alias{make.formulae_}
\alias{make.qtl.covar.model.df_}
\alias{make.response.model.df_}
\alias{remove.qtl.terms_}
\alias{replace.markers.with.add.dom_}
\alias{wrangle.genoprob.df_}
\alias{wrangle.loc.info.df_}
\alias{wrangle.scanonevar.input_}
\alias{wrangle.scanonevar.types_}
\title{wrangle.scanonevar.input_}
\usage{
wrangle.scanonevar.input_(cross, mean.formula, var.formula, chrs)

wrangle.scanonevar.types_(mean.formula, var.formula)

is.mean.formula(x)

is.var.formula(x)

is.formulae(x)

make.formulae_(mean.formula, var.formula)

replace.markers.with.add.dom_(cross, mean.formula, var.formula)

remove.qtl.terms_(formulae)

LOD(alt, null)

wrangle.loc.info.df_(cross, chrs = qtl::chrnames(cross))

wrangle.genoprob.df_(cross)

make.response.model.df_(cross, formulae)

make.qtl.covar.model.df_(cross, formulae)

additive.component(genoprobs.long)

dominance.component(genoprobs.long)
}
\arguments{
\item{cross}{The \code{cross}, built by \pkg{qtl} to be used in mapping}

\item{mean.formula}{a mean.formula}

\item{var.formula}{The formula to describe the residual variance of the
phenotype.  Keywords are var.QTL.add and var.QTL.dom for the additive and
dominance components of the QTL effect on residual phenotype variance.}

\item{x}{object to be tested whether it is a valid mean.formula}

\item{formulae}{The formulae from which QTL terms will be removed}

\item{alt}{alternative DGLM}

\item{null}{null DGLM}

\item{genoprobs.long}{The genoprobs from which the additive genetic
component should be extracted.  The 'long' format implies that each row
has information on the genoprob of one individual having one allele.}

\item{x}{object to be tested whether it is a valid var.formula}

\item{x}{object to be tested whether it is a valid formulae}

\item{var.forula}{a var.formula}

\item{genoprobs.long}{The genoprobs from which the additive genetic
component should be extracted.  The 'long' format implies that each row
has information on the genoprob of one individual having one allele.}
}
\value{
a list with three elements: 'model.df', the data.frame to be used
at every genetic locus in the upcoming scanonevar_, 'scan.df', the
data.frame that will be picked from, one genetic at a time to fill the
'QTL' variable in the model and 'scan.types', which indicates which types
of scan to do: mqtl, vqtl, and/or mvqtl.

either 'mean', 'var', or c('mean', 'var', 'joint'), indicating which
types of scans are implied by mean.formula and var.formula

TRUE if x is a valid mean.formula, FALSE otherwise

TRUE if X is a valid var.formula, FALSE otherwise

TRUE if x is a valid formulae, FALSE otherwise

a formulae, a list of length two where the first element is named
'mean.formula' and is a mean.formula and the second element is named
'var.formula' and is a var.formula.

A a list with two elements, mean.formula and var.formula,
where any markers from \code{cross} in the input mean.formula or var.formula
have been replaced with (marker.name_add + marker.name_dom).

a formulae object similar to the input formulae, but with QTL terms removed

log( likelihood of alternative model / likelihood of null model )

a tibble of the response in mean.formua

a tibble of the qtl keyword in formulae

A vector of the additive genetic component at the locus

A vector of the dominance genetic component at the locus
}
\examples{
x <- 27599

}

