% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_var_plots.R
\name{mean_var_plot_model_free}
\alias{mean_var_plot_model_based}
\alias{mean_var_plot_model_free}
\title{mean_var_plot_model_free}
\usage{
mean_var_plot_model_free(cross, phenotype.name, grouping.factor.names)

mean_var_plot_model_based(cross, phenotype.name, focal.groups = NULL,
  nuisance.groups = NULL, genotype.names = c("AA", "AB", "BB"),
  xlim = NULL, ylim = NULL, title = paste(phenotype.name, "by",
  paste(focal.groups, collapse = ", ")))
}
\arguments{
\item{cross}{the cross object from which phenotypes and genotypes are drawn}

\item{phenotype.name}{the name of the phenotype}

\item{grouping.factor.names}{the factors to group individuals by}

\item{focal.groups}{the focal covariates, whose effects will be plotted.  Markers or phenotypes.}

\item{nuisance.groups}{the nuisance covariates, whose effects will be modeled, then marginalized over.  Markers or phenotypes.}

\item{cross}{the cross}

\item{phenotype.name}{the name of the phenotype of interest}
}
\value{
Nothing, just plot.

nothing, just the plot.
}

