% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictive.plot.R
\name{predictive.plot}
\alias{predictive.plot}
\title{Plot Predictive Interval for Categorical Genotype/Phenotype Groups}
\usage{
predictive.plot(cross, mean.formula, var.formula, group.by,
  title = paste("Predictive Mean and Variance Confidence Intervals of",
  response.phen), xlim = NA, ylim = NA)
}
\arguments{
\item{cross}{The cross object to be plotted}

\item{mean.formula}{the mean formula}

\item{var.formula}{the var formula}

\item{title}{Optionally, title for the plot.  Defaults to 'Predictive of [response phenotype] from
[predictive phenotype (e.g. sex)] and [marker name]}

\item{xlim}{Optionally specify x-axis limits.  Defaults to data-dependent.}

\item{ylim}{Optionally specify y-axis limits.  Defaults to data.dependent.}
}
\value{
The plot
}
\description{
\code{predictive.plot} should be used to visually investigate loci identified
  with plot.scanonevar or summary.scanonevar.  The user can specify the same mean and variance
  formulae that were used in the scan, or specify new formulae to investigate interactions.
}
\examples{
  set.seed(27599)
  my.cross <- sim.cross(map = sim.map(), type = 'f2')
  my.cross <- calc.genoprob(my.cross)

}
\author{
Robert Corty \email{rcorty@gmail.com}
}

