% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanonevar.perm.R
\name{scanonevar.perm}
\alias{scanonevar.perm}
\title{Conduct Scanonevars on Permuted Genotype Data}
\usage{
scanonevar.perm(cross, mean.formula, var.formula, n.perms,
  chrs = unique(names(cross$geno)), verbose.return = FALSE)
}
\arguments{
\item{cross}{The \code{cross}, built by \pkg{qtl} to be used in mapping}

\item{mean.formula}{The formula to describe the mean of the phenotype.
Keywords are mean.QTL.add and mean.QTL.dom for the additive and dominance
components of the QTL effect on the mean.}

\item{var.formula}{The formula to describe the residual variance of the
phenotype.  Keywords are var.QTL.add and var.QTL.dom for the additive and
dominance components of the QTL effect on residual phenotype variance.}

\item{n.perms}{the number of permutations to conduct}
}
\value{
Returns a tbl_df of maximum LOD score observed in each genome scan for each chromosome type.
}
\description{
\code{scanonevar.perm} conducts many \code{scanonevar}s on permuted data
  and returns the maximum observed LOD score for each chromosome type in each scan.
  The results should be put into \code{convert.scanonevar.to.empirical.ps} with a scan in LODs
  to convert that scan to empirical p-values.  It's important that all the parameters used in
  \code{scanonevar.perm} are the same as the parameters that were used in the \code{scanonevar}
  that they will be used to convert to empirical ps.
}
\details{
It is recommended to use approximately 1000 permuted scans to produce highly-replicable,
  publication-quality empirial p-values.  For this purpose, users are recommended to dispatch this
  function to many computers in parallel, carefully setting the seed on each computer to insure
  pseudo-randomness.

none
}
\examples{
\dontrun{
  my.perms <- scanonevar.perm(cross = my.cross,
                              n.perms = 10))

}
}
\author{
Robert Corty \email{rcorty@gmail.com}
}
\seealso{
\code{\link{scanonevar}}, \code{\link{scanonevar.to.p.values}}
}

