% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanonevar_.R
\name{scanonevar_}
\alias{LOD}
\alias{additive.component}
\alias{dominance.component}
\alias{initialize.scanonevar.result_}
\alias{make.loc.specific.modeling.df}
\alias{scanonevar_}
\title{scanonevar_}
\usage{
scanonevar_(modeling.df, genoprob.df, loc.info.df, scan.types, scan.formulae,
  return.covar.effects)

initialize.scanonevar.result_(loc.info.df, scan.types, scan.formulae,
  return.covar.effects)

LOD(alt, null)

make.loc.specific.modeling.df(general.modeling.df, loc.genoprobs,
  model.formulae)

additive.component(genoprobs.long)

dominance.component(genoprobs.long)
}
\arguments{
\item{modeling.df}{the tbl_df that contains the response, all covariates,
and NA columns appropriate for marker effects}

\item{genoprob.df}{the tbl_df that contains the probability of each genotype
for each individual at each loc, where a 'loc' is either a marker or a pseudomarker}

\item{loc.info.df}{the tbl_df that contains information about the
locs that are being scanned, minimally a 'loc.name', 'chr' (short for
'chromosome') and 'pos' (short for 'position') for each loc.}

\item{scan.types}{Type(s) of scan(s) to be conducted, one of: 'mean',
'var', or c('mean', 'var', 'joint')}

\item{alt}{alternative DGLM}

\item{null}{null DGLM}

\item{general.modeling.df}{The modeling df that will be used across the scan,
should have NA for all QTL columns.  These columns will be filled in in the result.}

\item{loc.genoprobs}{The genoprobs of all individuals of all allels at this loc.}

\item{model.formulae}{The formulae used for the scan.}

\item{genoprobs.long}{The genoprobs from which the additive genetic
component should be extracted.  The 'long' format implies that each row
has information on the genoprob of one individual having one allele.}

\item{genoprobs.long}{The genoprobs from which the additive genetic
component should be extracted.  The 'long' format implies that each row
has information on the genoprob of one individual having one allele.}
}
\value{
a scanonevar object

A tbl_df with the metadata for a scanonevar, once filled in by
the scanonevar function, it will be a scanonevar object.

log( likelihood of alternative model / likelihood of null model)

a modeling df appropriate for modeling at one specific loc

A vector of the additive genetic component at the locus

A vector of the dominance genetic component at the locus
}
\details{
This internal function should not typically be called by a user.
}
\examples{
x <- 27599

x <- dplyr::data_frame(loc.name = c('m1', 'm2'),
                       chr = c(1, 1),
                       pos = c(5, 10))
initialize.scanonevar.result_(loc.info.df = x,
                              scan.types = 'mean')

x <- 27599
}
\author{
Robert W. Corty
}

