% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanonevar.R
\name{summary.scanonevar}
\alias{summary.scanonevar}
\title{summary.scanonevar}
\usage{
\method{summary}{scanonevar}(object, units = c("lod", "asymp.p",
  "empir.p"), thresh, ...)
}
\arguments{
\item{object}{the scanonevar object to be summarized}

\item{units}{Which units should be used to summarise?  'lod', 'asymp.p', or 'empir.p'}

\item{thresh}{the threshold over which (for LODs) or under which (for emprirical p values)
a locus will be printed.}

\item{...}{additional arguments controlling the summary}
}
\value{
None.  Only prints results to screen.
}
\description{
\code{summary.scanonevar} prints out the loci in a scanonevar object
  that exceed \code{thresh}.  It is an S3 generic for summary().  It handles scanonevar
  objects in both LOD units and empirical p value units.
}
\details{
none
}
\author{
Robert Corty \email{rcorty@gmail.com}
}
