% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.scanonevar.R
\name{summary.scanonevar}
\alias{summary.scanonevar}
\title{summary.scannevar}
\usage{
\method{summary}{scanonevar}(object, units = c("LOD", "empir.p"), thresh, ...)
}
\arguments{
\item{object}{the scanonevar to summarize}

\item{units}{the units}

\item{thresh}{the threshold}

\item{...}{ignored}
}
\value{
a summary of the scanonevar object
}

