% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/units.R
\name{units}
\alias{units}
\title{Get units of a scanonevar object}
\usage{
units(object)
}
\arguments{
\item{object}{the scanonevar object whose units are interrogated}
}
\value{
units of scanonevar object
}
\description{
Utility function to get the units of a scanonevar, which is either 'LODs'
   (logarithm of the odds between null and alternative model) or 'emp.ps' (empirically
   determined p.value of that LOD)
}
\details{


none
}
\author{
Robert Corty \email{rcorty@gmail.com}
}

