% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_validators.R
\name{is.scanonevar}
\alias{is.cross}
\alias{is.scanonevar}
\alias{is.scanonevar.w.perms}
\title{is.scanonevar}
\usage{
is.scanonevar(x)

is.scanonevar.w.perms(x)

is.cross(x)
}
\arguments{
\item{x}{The object being tested for whether or not is is a scanonevar.}

\item{x}{object being tested whether it is a scanonevar with perms}

\item{x}{The object being tested for whether or not is is a cross.}
}
\value{
TRUE is X is a scanonevar object, FALSE otherwise.

TRUE if x is a scanone var with perms (typically,
outputted from scanonevar.perm), and FALSE otherwise.

TRUE if x is a cross object, FALSE otherwise.
}
\examples{
is.scanonevar(x = 3)

test.cross <- qtl::sim.cross(map = qtl::sim.map(len = rep(20, 4), n.mar = 5))
test.cross <- qtl::calc.genoprob(cross = test.cross, step = 2)

x <- scanonevar(cross = test.cross)
is.scanonevar(x)

is.cross(3)
is.cross(qtl::sim.cross(map = qtl::sim.map()))

}
\author{
Robert W. Corty \email{rcorty@gmail.com}

Robert W. Corty \email{rcorty@gmail.com}
}

