% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_validate.input.scanonevar.R
\name{validate.input.scanonevar}
\alias{validate.input.scanonevar}
\title{Validate the Input before Conducting a Scanonevar}
\usage{
validate.input.scanonevar(cross, mean.formula, var.formula,
  chrs = names(cross$geno))
}
\arguments{
\item{cross}{The cross on which scanonevar will be performed, to be validated}

\item{mean.formula}{The formula to describe the mean of the phenotype.
Keywords are mean.QTL.add and mean.QTL.dom for the additive and dominance
components of the QTL effect on the mean.}

\item{var.formula}{The formula to describe the residual variance of the
phenotype.  Keywords are var.QTL.add and var.QTL.dom for the additive and
dominance components of the QTL effect on residual phenotype variance.}
}
\value{
Validated and organized inputs for scanonevar
}
\description{
\code{validate.input.scanonevar} should not typically be called by
  a user.  It validates the input to the \code{scanonevar} function.
}
\details{
none
}
\author{
Robert Corty \email{rcorty@gmail.com}
}

