% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrangle.scanonevar.input_.R
\name{wrangle.scanonevar.input_}
\alias{wrangle.genoprob.df_}
\alias{wrangle.loc.info.df_}
\alias{wrangle.modeling.df_}
\alias{wrangle.scan.formulae_}
\alias{wrangle.scan.types_}
\alias{wrangle.scanonevar.input_}
\title{wrangle.scanonevar.input_}
\usage{
wrangle.scanonevar.input_(cross, mean.formula, var.formula)

wrangle.scan.types_(mean.formula, var.formula)

wrangle.scan.formulae_(mean.formula, var.formula)

wrangle.loc.info.df_(cross)

wrangle.genoprob.df_(cross)

wrangle.modeling.df_(cross, scan.formulae)
}
\arguments{
\item{cross}{The \code{cross}, built by \pkg{qtl} to be used in mapping}

\item{mean.formula}{The formula to describe the mean of the phenotype.
Keywords are mean.QTL.add and mean.QTL.dom for the additive and dominance
components of the QTL effect on the mean.}

\item{var.formula}{The formula to describe the residual variance of the
phenotype.  Keywords are var.QTL.add and var.QTL.dom for the additive and
dominance components of the QTL effect on residual phenotype variance.}
}
\value{
a list with three elements: 'model.df', the data.frame to be used
at every genetic locus in the upcoming scanonevar_, 'scan.df', the
data.frame that will be picked from, one genetic at a time to fill the
'QTL' variable in the model and 'scan.types', which indicates which types
of scan to do: mqtl, vqtl, and/or mvqtl.
}
\examples{
x <- 27599

}
\author{
Robert W. Corty \email{rcorty@gmail.com}
}

